Refactor remote app-server namespaces and folders
