+++ src/JKToolKit.CodexSDK/AppServer/Remote/Internal/Infrastructure/RemoteStartMetadata.cs
@@ -1,4 +1,4 @@
-namespace JKToolKit.CodexSDK.AppServer;
+namespace JKToolKit.CodexSDK.AppServer.Remote.Internal;
 
 internal sealed record RemoteStartMetadata(
     string Id,
