+++ src/JKToolKit.CodexSDK/AppServer/Remote/Internal/Orchestration/RemoteAppServerConnector.cs
@@ -1,4 +1,7 @@
-namespace JKToolKit.CodexSDK.AppServer;
+using JKToolKit.CodexSDK.AppServer;
+using JKToolKit.CodexSDK.AppServer.Remote;
+
+namespace JKToolKit.CodexSDK.AppServer.Remote.Internal;
 
 internal sealed class RemoteAppServerConnector
 {
