+++ src/JKToolKit.CodexSDK/AppServer/Remote/Models/CodexRemoteAppServerEntry.cs
@@ -1,4 +1,4 @@
-namespace JKToolKit.CodexSDK.AppServer;
+namespace JKToolKit.CodexSDK.AppServer.Remote;
 
 /// <summary>
 /// Registry entry for a SDK-managed remote Codex app-server process.
