+++ tests/JKToolKit.CodexSDK.Tests/Unit/RemoteAppServerManagerTests.cs
@@ -2,6 +2,9 @@ using System.Text.Json;
 using FluentAssertions;
 using JKToolKit.CodexSDK.AppServer;
 using JKToolKit.CodexSDK.AppServer.Internal;
+using JKToolKit.CodexSDK.AppServer.Remote;
+using JKToolKit.CodexSDK.AppServer.Remote.Internal;
+using JKToolKit.CodexSDK.AppServer.Remote.Registry;
 using JKToolKit.CodexSDK.Exec;
 using JKToolKit.CodexSDK.Infrastructure.JsonRpc;
 using JKToolKit.CodexSDK.Infrastructure.JsonRpc.Messages;
